requirements.txt
Flask==2.3.3
flask-restful==0.3.10
flask-jwt-extended==4.5.2
pymongo==4.5.0
bcrypt==4.0.1
python-dotenv==1.0.0
Werkzeug==2.3.7
numpy==2.2.4
pandas==2.2.3
openpyxl==3.1.2
requests==2.31.0 